Add kiosk CLI module
