+++ pyproject.toml
@@ -35,7 +35,7 @@ piwardrive-prefetch-batch = "piwardrive.scripts.prefetch_batch:main"
 log-follow = "piwardrive.scripts.log_follow:main"
 config-cli = "piwardrive.scripts.config_cli:main"
 calibrate-orientation = "piwardrive.scripts.calibrate_orientation:main"
-piwardrive-kiosk = "piwardrive.scripts.kiosk:main"
+piwardrive-kiosk = "piwardrive.cli.kiosk:main"
 export-log-bundle = "piwardrive.scripts.export_log_bundle:main"
