+++ src/piwardrive/kiosk.py
@@ -1,0 +1,2 @@
+"""Backward-compatible entry for the kiosk CLI."""
+from .cli.kiosk import *  # noqa: F401,F403
